+++ sources/packages/backend/apps/queue-consumers/src/processors/schedulers/institution-integration/ier12-integration/models/ier.model.ts
@@ -1,8 +1,3 @@
-export class IER12ResultQueueOutDTO {
-  generatedFile: string;
-  uploadedRecords: number;
-}
-
 export class GeneratedDateQueueInDTO {
   generatedDate?: string;
 }
