+++ sources/packages/backend/libs/integrations/src/atbc-integration/atbc-integration.module.ts
@@ -1,6 +1,6 @@
 import { Module } from "@nestjs/common";
 import { SystemUserModule } from "@sims/services/system-users";
-import { ATBCIntegrationProcessingService } from "./atbc-integration-processing.service";
+import { ATBCIntegrationProcessingService } from "./atbc-integration.processing.service";
 import { StudentService, ATBCService } from "../services";
 
 @Module({
