+++ sources/packages/backend/libs/integrations/src/atbc-integration/index.ts
@@ -1,2 +1,2 @@
 export * from "./atbc-integration.module";
-export * from "./atbc-integration-processing.service";
+export * from "./atbc-integration.processing.service";
