+++ sources/packages/backend/libs/integrations/src/cra-integration/cra-personal-verification.service.ts
@@ -16,7 +16,7 @@ import {
 import { getUTCNow } from "@sims/utilities";
 import * as path from "path";
 import { ConfigService } from "@sims/utilities/config";
-import { CRAIntegrationService } from "./cra-integration.service";
+import { CRAIntegrationService } from "./cra.integration.service";
 import { CRAIncomeVerificationsService } from "../services";
 
 const INCOME_VERIFICATION_TAG = "VERIFICATION_ID";
