+++ sources/packages/backend/libs/integrations/src/esdc-integration/disbursement-receipt-integration/disbursement-receipt-request.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from "@nestjs/common";
 import { InjectLogger, LoggerService } from "@sims/utilities/logger";
-import { DisbursementReceiptIntegrationService } from "./disbursement-receipt-integration.service";
+import { DisbursementReceiptIntegrationService } from "./disbursement-receipt.integration.service";
 import { DAILY_DISBURSEMENT_REPORT_NAME } from "@sims/services/constants";
 import { DisbursementReceiptService } from "@sims/integrations/services";
 import { ReportService, ReportsFilterModel } from "@sims/services";
