+++ sources/packages/backend/libs/integrations/src/esdc-integration/e-cert-integration/e-cert-part-time-integration/e-cert-part-time.integration.service.ts
@@ -1,0 +1,161 @@
+import { Injectable } from "@nestjs/common";
+import {
+  RecordTypeCodes,
+  CSGD,
+  CSGP,
+  CSGPT,
+  Award,
+  ECertRecord,
+} from "../models/e-cert-integration-model";
+import { ECertPartTimeFileHeader } from "./e-cert-files/e-cert-file-header";
+import { ECertPartTimeFileFooter } from "./e-cert-files/e-cert-file-footer";
+import { ECertPartTimeFileRecord } from "./e-cert-files/e-cert-file-record";
+import { DisbursementValueType, OfferingIntensity } from "@sims/sims-db";
+import { ECertIntegrationService } from "../e-cert.integration.service";
+import { ECertResponseRecord } from "../e-cert-files/e-cert-response-record";
+import { ConfigService } from "@sims/utilities/config";
+import { FixedFormatFileLine } from "@sims/integrations/services/ssh";
+import { SshService } from "@sims/integrations/services";
+import {
+  combineDecimalPlaces,
+  getDisbursementAmountByValueCode,
+  getGenderCode,
+  getPartTimeMaritalStatusCode,
+  getTotalDisbursementAmount,
+  round,
+} from "@sims/utilities";
+
+/**
+ * Manages the file content generation and methods to
+ * upload/download the files to SFTP.
+ */
+@Injectable()
+export class ECertPartTimeIntegrationService extends ECertIntegrationService {
+  constructor(
+    private readonly eCertPartTimeFileHeader: ECertPartTimeFileHeader,
+    private readonly eCertPartTimeFileFooter: ECertPartTimeFileFooter,
+    config: ConfigService,
+    sshService: SshService,
+  ) {
+    super(config.zoneBSFTP, sshService);
+  }
+
+  /**
+   * Create the ECert file content, by populating the
+   * header, detail and trailer records.
+   * @param ecertRecords student, User and application data.
+   * @returns complete ECert content to be sent.
+   */
+  createRequestContent(ecertRecords: ECertRecord[]): FixedFormatFileLine[] {
+    const fileLines: FixedFormatFileLine[] = [];
+    // Header record
+    const header = new ECertPartTimeFileHeader();
+    header.recordTypeCode = RecordTypeCodes.ECertPartTimeHeader;
+    header.processDate = new Date();
+    fileLines.push(header);
+
+    // Detail records
+    // Calculated values
+    const fileRecords = ecertRecords.map((ecertRecord) => {
+      // ! All dollar values must be rounded to the nearest integer (0.5 rounds up)
+      const roundedAwards = ecertRecord.awards.map(
+        (award) =>
+          ({
+            valueType: award.valueType,
+            valueCode: award.valueCode,
+            valueAmount: round(award.valueAmount).toString(),
+          } as Award),
+      );
+
+      const totalCSGPPTAmount = getDisbursementAmountByValueCode(
+        roundedAwards,
+        CSGD,
+      );
+      const totalCSGPPDAmount = getDisbursementAmountByValueCode(
+        roundedAwards,
+        CSGP,
+      );
+      const totalCSGPPTDEPAmount = getDisbursementAmountByValueCode(
+        roundedAwards,
+        CSGPT,
+      );
+
+      const disbursementAmount = getTotalDisbursementAmount(roundedAwards, [
+        DisbursementValueType.CanadaLoan,
+      ]);
+      const totalGrantAmount = getTotalDisbursementAmount(roundedAwards, [
+        DisbursementValueType.CanadaGrant,
+        DisbursementValueType.BCTotalGrant,
+      ]);
+      const totalBCSGAmount = getTotalDisbursementAmount(roundedAwards, [
+        DisbursementValueType.BCTotalGrant,
+      ]);
+
+      const record = new ECertPartTimeFileRecord();
+      record.recordType = RecordTypeCodes.ECertPartTimeRecord;
+      record.sin = ecertRecord.sin;
+      record.courseLoad = ecertRecord.courseLoad;
+      record.certNumber = ecertRecord.documentNumber;
+      record.disbursementDate = ecertRecord.disbursementDate;
+      record.documentProducedDate = ecertRecord.documentProducedDate;
+      record.disbursementAmount = disbursementAmount;
+      record.schoolAmount = combineDecimalPlaces(ecertRecord.schoolAmount);
+      record.educationalStartDate = ecertRecord.educationalStartDate;
+      record.educationalEndDate = ecertRecord.educationalEndDate;
+      record.federalInstitutionCode = ecertRecord.federalInstitutionCode;
+      record.weeksOfStudy = ecertRecord.weeksOfStudy;
+      record.fieldOfStudy = ecertRecord.fieldOfStudy;
+      record.yearOfStudy = ecertRecord.yearOfStudy;
+      record.enrollmentConfirmationDate =
+        ecertRecord.enrollmentConfirmationDate;
+      record.dateOfBirth = ecertRecord.dateOfBirth;
+      record.lastName = ecertRecord.lastName;
+      record.firstName = ecertRecord.firstName;
+      record.addressLine1 = ecertRecord.addressLine1;
+      record.addressLine2 = ecertRecord.addressLine2;
+      record.city = ecertRecord.city;
+      record.emailAddress = ecertRecord.email;
+      record.gender = getGenderCode(ecertRecord.gender);
+      record.maritalStatus = getPartTimeMaritalStatusCode(
+        ecertRecord.maritalStatus,
+      );
+      record.studentNumber = ecertRecord.studentNumber;
+      record.totalGrantAmount = totalGrantAmount;
+      record.totalBCSGAmount = totalBCSGAmount;
+      record.totalCSGPPTAmount = totalCSGPPTAmount;
+      record.totalCSGPPDAmount = totalCSGPPDAmount;
+      record.totalCSGPPTDEPAmount = totalCSGPPTDEPAmount;
+      return record;
+    });
+    fileLines.push(...fileRecords);
+
+    // Footer or Trailer record
+
+    // Total disbursementAmount disbursed in the disbursement file.
+    const totalAmountDisbursed = fileRecords.reduce(
+      (hash, record) => hash + +record.disbursementAmount,
+      0,
+    );
+    const footer = new ECertPartTimeFileFooter();
+    footer.recordTypeCode = RecordTypeCodes.ECertPartTimeFooter;
+    footer.totalAmountDisbursed = totalAmountDisbursed;
+    footer.recordCount = fileRecords.length;
+    fileLines.push(footer);
+
+    return fileLines;
+  }
+
+  /**
+   * This method will call the appropriate common implementation by passing the appropriate parameters.
+   * @param remoteFilePath E-Cert response file to be processed.
+   * @returns Parsed records from the file.
+   */
+  downloadResponseFile(remoteFilePath: string): Promise<ECertResponseRecord[]> {
+    return this.downloadECertResponseFile(
+      remoteFilePath,
+      this.eCertPartTimeFileHeader,
+      this.eCertPartTimeFileFooter,
+      OfferingIntensity.partTime,
+    );
+  }
+}
