+++ sources/packages/backend/libs/integrations/src/esdc-integration/index.ts
@@ -2,9 +2,9 @@ export * from "./disbursement-receipt-integration/disbursement-receipt-files/dis
 export * from "./disbursement-receipt-integration/disbursement-receipt-files/disbursement-receipt-file-footer";
 export * from "./disbursement-receipt-integration/disbursement-receipt-files/disbursement-receipt-file-header";
 export * from "./disbursement-receipt-integration/disbursement-receipt-files/disbursement-receipt-file-record";
-export * from "./disbursement-receipt-integration/disbursement-receipt-integration.service";
+export * from "./disbursement-receipt-integration/disbursement-receipt.integration.service";
 export * from "./disbursement-receipt-integration/disbursement-receipt-integration.module";
-export * from "./disbursement-receipt-integration/disbursement-receipt-processing.service";
+export * from "./disbursement-receipt-integration/disbursement-receipt.processing.service";
 export * from "./disbursement-receipt-integration/disbursement-receipt-request.service";
 export * from "./disbursement-receipt-integration/models/disbursement-receipt-integration.model";
 export * from "./e-cert-integration/e-cert-file-handler";
@@ -16,20 +16,20 @@ export * from "./e-cert-integration/e-cert-full-time-integration/e-cert-files/e-
 export * from "./e-cert-integration/e-cert-full-time-integration/e-cert-files/e-cert-file-header";
 export * from "./e-cert-integration/e-cert-full-time-integration/e-cert-files/e-cert-file-record";
 export * from "./e-cert-integration/e-cert-full-time-integration/e-cert-files/e-cert-response-record";
-export * from "./e-cert-integration/e-cert-full-time-integration/e-cert-full-time-integration.service";
+export * from "./e-cert-integration/e-cert-full-time-integration/e-cert-full-time.integration.service";
 export * from "./e-cert-integration/e-cert-integration.module";
-export * from "./e-cert-integration/e-cert-integration.service";
+export * from "./e-cert-integration/e-cert.integration.service";
 export * from "./e-cert-integration/e-cert-part-time-integration/e-cert-files/e-cert-file-footer";
 export * from "./e-cert-integration/e-cert-part-time-integration/e-cert-files/e-cert-file-header";
 export * from "./e-cert-integration/e-cert-part-time-integration/e-cert-files/e-cert-file-record";
 export * from "./e-cert-integration/e-cert-part-time-integration/e-cert-files/e-cert-response-record";
-export * from "./e-cert-integration/e-cert-part-time-integration/e-cert-part-time-integration.service";
+export * from "./e-cert-integration/e-cert-part-time-integration/e-cert-part-time.integration.service";
 export * from "./e-cert-integration/e-cert-response-record-identification";
 export * from "./esdc-file-handler";
 export * from "./fed-restriction-integration/fed-restriction-files/fed-restriction-file-record";
 export * from "./fed-restriction-integration/fed-restriction-integration.module";
-export * from "./fed-restriction-integration/fed-restriction-integration.service";
-export * from "./fed-restriction-integration/fed-restriction-processing.service";
+export * from "./fed-restriction-integration/fed-restriction.integration.service";
+export * from "./fed-restriction-integration/fed-restriction.processing.service";
 export * from "./msfaa-integration/msfaa-files/msfaa-file-detail";
 export * from "./msfaa-integration/msfaa-files/msfaa-file-footer";
 export * from "./msfaa-integration/msfaa-files/msfaa-file-header";
@@ -44,6 +44,6 @@ export * from "./sin-validation/sin-validation-files/sin-validation-file-footer"
 export * from "./sin-validation/sin-validation-files/sin-validation-file-header";
 export * from "./sin-validation/sin-validation-files/sin-validation-file-request";
 export * from "./sin-validation/sin-validation-files/sin-validation-file-response";
-export * from "./sin-validation/sin-validation-integration.service";
-export * from "./sin-validation/sin-validation-processing.service";
+export * from "./sin-validation/sin-validation.integration.service";
+export * from "./sin-validation/sin-validation.processing.service";
 export * from "./sin-validation/sin-validation.module";
