+++ sources/packages/backend/libs/integrations/src/esdc-integration/sin-validation/sin-validation.module.ts
@@ -1,7 +1,7 @@
 import { Module } from "@nestjs/common";
 import { SequenceControlService } from "@sims/services";
-import { SINValidationIntegrationService } from "./sin-validation-integration.service";
-import { SINValidationProcessingService } from "./sin-validation-processing.service";
+import { SINValidationIntegrationService } from "./sin-validation.integration.service";
+import { SINValidationProcessingService } from "./sin-validation.processing.service";
 import { ConfigModule } from "@sims/utilities/config";
 import {
   StudentService,
