+++ sources/packages/backend/libs/integrations/src/institution-integration/ece-integration/ece-file.service.ts
@@ -8,7 +8,7 @@ import {
 import { getFileNameAsCurrentTimestamp } from "@sims/utilities";
 import { DisbursementScheduleService } from "@sims/integrations/services";
 import { ECERecord, ECEUploadResult } from "./models/ece-integration.model";
-import { ECEIntegrationService } from "./ece-integration.service";
+import { ECEIntegrationService } from "./ece.integration.service";
 
 @Injectable()
 export class ECEFileService {
