+++ sources/packages/backend/libs/integrations/src/institution-integration/ece-integration/index.ts
@@ -1,5 +1,5 @@
 export * from "./models/ece-integration.model";
 export * from "./ece-files/ece-file-detail";
 export * from "./ece-file.service";
-export * from "./ece-integration.service";
+export * from "./ece.integration.service";
 export * from "./ece-integration.module";
