+++ sources/packages/backend/libs/integrations/src/institution-integration/ier12-integration/ier12-file.service.ts
@@ -6,7 +6,7 @@ import {
   InstitutionIntegrationConfig,
 } from "@sims/utilities/config";
 import { getFileNameAsCurrentTimestamp } from "@sims/utilities";
-import { IER12IntegrationService } from "./ier12-integration.service";
+import { IER12IntegrationService } from "./ier12.integration.service";
 import {
   IER12Record,
   IER12UploadResult,
