+++ sources/packages/backend/libs/integrations/src/institution-integration/ier12-integration/ier12-integration.module.ts
@@ -1,7 +1,7 @@
 import { Global, Module } from "@nestjs/common";
 import { WorkflowClientService } from "@sims/services";
 import { ConfigModule } from "@sims/utilities/config";
-import { IER12IntegrationService } from "./ier12-integration.service";
+import { IER12IntegrationService } from "./ier12.integration.service";
 import { IER12FileService } from "./ier12-file.service";
 import {
   StudentAssessmentService,
