+++ sources/packages/backend/libs/integrations/src/institution-integration/ier12-integration/index.ts
@@ -1,5 +1,5 @@
 export * from "./models/ier12-integration.model";
 export * from "./ier12-file-detail";
 export * from "./ier12-file.service";
-export * from "./ier12-integration.service";
+export * from "./ier12.integration.service";
 export * from "./ier12-integration.module";
