+++ sources/packages/backend/libs/integrations/src/sfas-integration/index.ts
@@ -1,2 +1,2 @@
 export * from "./sfas-integration.module";
-export * from "./sfas-integration-processing.service";
+export * from "./sfas-integration.processing.service";
